Move color conversion functions to utils
